Initialize IFontRenderData & DxFontRenderData
