+++ src/renderer/dx/DxRenderer.hpp
@@ -153,16 +153,6 @@ namespace Microsoft::Console::Render
         bool _isEnabled;
         bool _isPainting;
 
-        struct LineMetrics
-        {
-            float gridlineWidth;
-            float underlineOffset;
-            float underlineOffset2;
-            float underlineWidth;
-            float strikethroughOffset;
-            float strikethroughWidth;
-        };
-
         LineMetrics _lineMetrics;
         til::size _displaySizePixels;
         til::size _glyphCell;
@@ -195,11 +185,6 @@ namespace Microsoft::Console::Render
         ::Microsoft::WRL::ComPtr<ID2D1Factory1> _d2dFactory;
 
         ::Microsoft::WRL::ComPtr<IDWriteFactory1> _dwriteFactory;
-        ::Microsoft::WRL::ComPtr<IDWriteTextFormat> _dwriteTextFormat;
-        ::Microsoft::WRL::ComPtr<IDWriteTextFormat> _dwriteTextFormatItalic;
-        ::Microsoft::WRL::ComPtr<IDWriteFontFace1> _dwriteFontFace;
-        ::Microsoft::WRL::ComPtr<IDWriteFontFace1> _dwriteFontFaceItalic;
-        ::Microsoft::WRL::ComPtr<IDWriteTextAnalyzer1> _dwriteTextAnalyzer;
         ::Microsoft::WRL::ComPtr<CustomTextLayout> _customLayout;
         ::Microsoft::WRL::ComPtr<CustomTextRenderer> _customRenderer;
         ::Microsoft::WRL::ComPtr<ID2D1StrokeStyle> _strokeStyle;
@@ -299,33 +284,6 @@ namespace Microsoft::Console::Render
         [[nodiscard]] HRESULT _CopyFrontToBack() noexcept;
 
         [[nodiscard]] HRESULT _EnableDisplayAccess(const bool outputEnabled) noexcept;
-
-        [[nodiscard]] ::Microsoft::WRL::ComPtr<IDWriteFontFace1> _ResolveFontFaceWithFallback(std::wstring& familyName,
-                                                                                              DWRITE_FONT_WEIGHT& weight,
-                                                                                              DWRITE_FONT_STRETCH& stretch,
-                                                                                              DWRITE_FONT_STYLE& style,
-                                                                                              std::wstring& localeName) const;
-
-        [[nodiscard]] ::Microsoft::WRL::ComPtr<IDWriteFontFace1> _FindFontFace(std::wstring& familyName,
-                                                                               DWRITE_FONT_WEIGHT& weight,
-                                                                               DWRITE_FONT_STRETCH& stretch,
-                                                                               DWRITE_FONT_STYLE& style,
-                                                                               std::wstring& localeName) const;
-
-        [[nodiscard]] std::wstring _GetLocaleName() const;
-
-        [[nodiscard]] std::wstring _GetFontFamilyName(gsl::not_null<IDWriteFontFamily*> const fontFamily,
-                                                      std::wstring& localeName) const;
-
-        [[nodiscard]] HRESULT _GetProposedFont(const FontInfoDesired& desired,
-                                               FontInfo& actual,
-                                               const int dpi,
-                                               ::Microsoft::WRL::ComPtr<IDWriteTextFormat>& textFormat,
-                                               ::Microsoft::WRL::ComPtr<IDWriteTextFormat>& textFormatItalic,
-                                               ::Microsoft::WRL::ComPtr<IDWriteTextAnalyzer1>& textAnalyzer,
-                                               ::Microsoft::WRL::ComPtr<IDWriteFontFace1>& fontFace,
-                                               ::Microsoft::WRL::ComPtr<IDWriteFontFace1>& fontFaceItalic,
-                                               LineMetrics& lineMetrics) const noexcept;
 
         [[nodiscard]] til::size _GetClientSize() const;
